\encoding{UTF-8}

\name{pop.wolf-package}
\alias{pop.wolf-package}
\alias{pop.wolf}
\docType{package}
\title{
Population models
}
\description{
A package to run simulations of wolf populations using an Individual-Based Model compiled in C. The model is parameterized with data from the Scandinavian wolf population.
}
\details{
\tabular{ll}{
Package: \tab pop.wolf\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-12-23\cr
License: \tab GPL-3\cr
}
}
\author{
Guillaume Chapron <guillaume.chapron@slu.se> with contributions from Camilla Wikenros, Olof Liberg, Øystein Flagstad, Cyril Milleret, Johan Månsson, Linn Svensson, Barbara Zimmermann, Mikael Åkesson, Petter Wabakken and Håkan Sand.
}

